+++ src/pas/plugins/authomatic/plugin.py
@@ -285,7 +285,9 @@ class AuthomaticPlugin(BasePlugin):
             if not userid_matches:
                 sheet = identity.propertysheet
                 fullname = sheet.getProperty('fullname')
-                if not fullname or fullname.encode("utf8").lower().find(search_id.lower()) == -1:
+                if six.PY2 and isinstance(fullname, six.text_type):
+                    fullname = fullname.encode('utf8')
+                if not fullname or fullname.lower().find(search_id.lower()) == -1:
                     continue
 
             identity_userid = identity.userid